requirements.txt
# HKJC Win Odds Trends Extraction & Upload System
# Essential Python packages for 獨贏賠率走勢 (Win Odds Trends) extraction

# Core web scraping and automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Database integration
pocketbase>=0.8.0

# Configuration management
python-dotenv>=1.0.0

# Data processing (optional but recommended)
pandas>=2.0.0

# Additional dependencies (automatically installed with above packages)
# certifi - SSL certificates
# charset-normalizer - Character encoding detection
# idna - Internationalized domain names
# urllib3 - HTTP client
# soupsieve - CSS selectors for BeautifulSoup
# pytz - Timezone handling
# python-dateutil - Date parsing utilities
